pyproject.toml
[project]
name = "clip-creator"
version = "0.1.0"
description = "Extract the best short clips from podcast episodes"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20231117",
    "openai>=1.14",
    "anthropic>=0.49",
    "librosa>=0.10",
    "soundfile>=0.12",
    "numpy>=1.26",
    "scipy>=1.12",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.scripts]
clip-creator = "clip_creator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
